LayoutTests/http/tests/site-isolation/pageswap-event-expected.txt

Tests that pageswap is fired when site isolation is enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.data.name is "pageswap"
PASS event.data.activation is null
PASS event.data.viewTransition is null
PASS successfullyParsed is true

TEST COMPLETE
